sql_php_phyton/projeto_login/backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
